Exchange with Lidia:
My branch with the changes is here:

Link: https://github.com/LidiaCucurull-NOAA/GSI/tree/ro_fix

Since you didn’t implement the Jacobian and stricter gross error QCs
from NCEP, you should only focus on a couple of changes in
setupbend.f90:

1- Reassignment of the location of an observation within the vertical
grid if the location is aliased to the lower section of the
non-monotonic grid
2- Reassignment of the xj values within the vertical grid if these
values are aliased to the lower section of the non-monotonic grid
3- Increase from “2” to “5” (this is a good value for us but it should
go higher as the number of vertical levels increase) in the
super-refraction QC when an observation is above a super-refraction
layer.

In principle, these changes should not affect biases.“1” and “2” are bug
fixes that didn’t show up with 64 levels and need to be implemented.
Number “3” is related to the limitations of the current forward operator
and they should go away with the new forward operator that Jim and I are
working on. (I will keep you posted on the progress of this work).

Let me know if you have any questions!

Have a nice evening,
Lidia