Update gmao_global_convinfo.txt
The 562 entry was only needed for identifying drifting buoys in MERRA2.  The read_prepbufr code was subsequently modified by NCEP so that the T29 value was no longer used for subtype identification so the code for changing the drifter KX was added explicitly to read_prepbufr.  Thus the 562 entry is obsolete and should be removed.