Delete unneeded call to closbf
The "closbf" call used to be needed because there was an "openbf" call earlier in the program. Between CVS v1.26.2.1 and v1.26.2.2 that "openbf" call was removed but the (now unneeded) "closbf" call was not removed.  It isn't a problem with the static allocation BUFR lib but it caused a problem when the BUFR lib was compiled with DYNAMIC allocation.   Let's just get rid of this call.